Move backup and sync settings to dedicated tab
