+++ trace/Views/Settings/GeneralSettingsView.swift
@@ -7,8 +7,6 @@
 
 import SwiftUI
 import Carbon
-import UniformTypeIdentifiers
-import os.log
 
 struct GeneralSettingsView: View {
     @Binding var launchAtLogin: Bool
@@ -20,13 +18,6 @@ struct GeneralSettingsView: View {
     @State private var hoveredAccent: TraceAccent?
     @State private var eventMonitor: Any?
     @State private var showingRestartAlert = false
-    @State private var showingExportSuccessAlert = false
-    @State private var showingImportAlert = false
-    @State private var showingImportFileImporter = false
-    @State private var showingExportFileExporter = false
-    @State private var exportedFileURL: URL?
-    @State private var importOverwriteExisting = false
-    @State private var importError: String?
     @ObservedObject private var settingsManager = SettingsManager.shared
     @Environment(\.colorScheme) private var colorScheme
     @Environment(\.traceTheme) private var traceTheme
@@ -271,33 +262,6 @@ struct GeneralSettingsView: View {
                 Text("Choose which result types appear in Trace search. Calendar search is opt-in and requires permission.")
             }
             
-            NativeSettingsSection("Settings Backup") {
-                NativeSettingsRow(
-                    title: "Export Settings",
-                    subtitle: "Save all your Trace settings to a file"
-                ) {
-                    Button("Export...") {
-                        exportSettings()
-                    }
-                    .buttonStyle(.bordered)
-                }
-                
-                NativeSettingsDivider()
-                
-                NativeSettingsRow(
-                    title: "Import Settings",
-                    subtitle: "Restore settings from a previously exported file"
-                ) {
-                    Button("Import...") {
-                        showingImportAlert = true
-                    }
-                    .buttonStyle(.bordered)
-                }
-            } footer: {
-                Text("Export includes hotkeys, custom folders, app preferences, and usage statistics.")
-            }
-
-            RemoteSettingsSyncView()
         }
         .onAppear {
             // Load settings from SettingsManager
@@ -318,111 +282,6 @@ struct GeneralSettingsView: View {
             Button("Later", role: .cancel) { }
         } message: {
             Text("The global hotkey change requires restarting Trace to take effect.")
-        }
-        .alert("Export Successful", isPresented: $showingExportSuccessAlert) {
-            Button("Open in Finder") {
-                if let url = exportedFileURL {
-                    NSWorkspace.shared.selectFile(url.path, inFileViewerRootedAtPath: "")
-                }
-            }
-            Button("OK") { }
-        } message: {
-            if let url = exportedFileURL {
-                Text("Settings exported to:\n\(url.lastPathComponent)")
-            }
-        }
-        .alert("Import Settings", isPresented: $showingImportAlert) {
-            Button("Replace All Settings") {
-                importOverwriteExisting = true
-                showingImportFileImporter = true
-            }
-            Button("Keep Existing (Add Missing)") {
-                importOverwriteExisting = false
-                showingImportFileImporter = true
-            }
-            Button("Cancel", role: .cancel) { }
-        } message: {
-            Text("Choose how to handle existing settings when importing.")
-        }
-        .fileImporter(
-            isPresented: $showingImportFileImporter,
-            allowedContentTypes: [UTType.json],
-            allowsMultipleSelection: false
-        ) { result in
-            handleImportFileSelection(result)
-        }
-    }
-    
-    // MARK: - Settings Import/Export
-    
-    private func exportSettings() {
-        do {
-            let fileURL = try settingsManager.exportToFile()
-            exportedFileURL = fileURL
-            showingExportSuccessAlert = true
-        } catch {
-            showAlert(title: "Export Failed", message: error.localizedDescription)
-        }
-    }
-    
-    private func handleImportFileSelection(_ result: Result<[URL], Error>) {
-        switch result {
-        case .success(let urls):
-            guard let url = urls.first else { return }
-            importSettings(from: url)
-        case .failure(let error):
-            showAlert(title: "File Selection Failed", message: error.localizedDescription)
-        }
-    }
-    
-    private func importSettings(from url: URL) {
-        do {
-            try settingsManager.importFromFile(url, overwriteExisting: importOverwriteExisting)
-            
-            // Reload all managers after import
-            reloadManagersAfterImport()
-            
-            showAlert(title: "Import Successful", 
-                     message: "Settings have been imported and all managers have been reloaded. All imported hotkeys should now work immediately.")
-        } catch {
-            showAlert(title: "Import Failed", message: error.localizedDescription)
-        }
-    }
-    
-    private func reloadManagersAfterImport() {
-        // Reload app hotkeys
-        AppHotkeyManager.shared.setupHotkeys()
-        
-        // Reload window hotkeys
-        let windowHotkeyManager = WindowHotkeyManager.shared
-        for (position, hotkeyData) in settingsManager.settings.windowHotkeys {
-            if let windowPosition = WindowPosition(rawValue: position) {
-                windowHotkeyManager.updateHotkey(
-                    for: windowPosition,
-                    keyCombo: hotkeyData.hotkey,
-                    keyCode: UInt32(hotkeyData.keyCode),
-                    modifiers: UInt32(hotkeyData.modifiers)
-                )
-            }
-        }
-        
-        // Reload quick links manager
-        DispatchQueue.main.async {
-            let serviceContainer = ServiceContainer.shared
-            serviceContainer.quickLinksManager.loadQuickLinks()
-        }
-
-        DictationHotkeyManager.shared.reload()
-    }
-    
-    private func showAlert(title: String, message: String) {
-        DispatchQueue.main.async {
-            let alert = NSAlert()
-            alert.messageText = title
-            alert.informativeText = message
-            alert.alertStyle = .informational
-            alert.addButton(withTitle: "OK")
-            alert.runModal()
         }
     }
